Task #108: completed prod distribution creation.

+++ deploy.sh
@@ -2,14 +2,32 @@
 
 if [ $# -eq 0 ]
   then
-    echo "Please provide the target environment [gorico_prod, gorico_dev, xxx_prod, xxx_dev]"
+    echo "Please provide the target application [gorico, xxx]"
     exit 0
 fi
 
-rm appdata.json
-ln -s $1.json appdata.json
+if [ ! -f ${1}_prod.json ]; then
+    echo "Target not found!"
+    exit 0
+fi
 
-#APINAME=`cat app_data.json | jq -r ".apiName"`
+#copy amplify.js into amplify.ts just in case any api has been added
+cp src/aws-exports.js src/aws-exports.ts
+
+#remove previous buils
+rm -rf dist/*
+
+rm appdata.json
+ln -s $1_prod.json appdata.json
+
+#re-deploy lambdas
+cd lambda 
+./deploy_lambdas.sh $1_prod
+
+#re-deploy dynamo-tables
+cd ../dynamo-tables
+./deploy_tables.sh $1_prod
+cd ..
 
 #compile application
 npm run build-prod
@@ -17,3 +35,7 @@ npm run build-prod
 aws s3 rm s3://gorico2.cloud/$1 --recursive
 #upload files
 aws s3 cp ./dist s3://gorico2.cloud/$1 --recursive --acl public-read
+
+#revert back to dev environment
+rm appdata.json
+ln -s $1_dev.json appdata.json
